Added math.sqrt, and updated functional example to use it.
We've added the math.sqrt function to the standard-library,
and have updated the `is_prime` example function to use it.

This means that when testing for a prime number we can test
the integers from 2->sqrt(N), rather than 2->N, which is
a good optimization to make.

(We add one, just to be safer.)