Allow +=, -=, /=, and *= to work.
This commit updates the parser and executor to allow the following
code to successfully execute:

        i = 10;
        i += 10;

Similarly "-=", "/=", and "*=" work.

We've implemented this functionality via a clever use (abuse?) of
the assignment operator, recording the operator which is used.  In
the case of assignment we just store "=", otherwise we store the
operator ("+=", "-="), so that we can correctly evaluate it.

Note that we've only implemented this for integers so far, we
could probably implement "+=" for strings at least.  But we probably
need double-support too.

This closes #14.