Improve implementation of constants, to obey scope.
We got into a mess in the previous series of commits because
we tried to store constants in a different map to variables,
then searching for them upwards - which meant that local variables
with the same name(s) would get missed.

To improve things such that constants and variables scope correctly
we've now given the notion of a "read-only" flag to EVERY object
type, and we set that only when `SetConst` is used.  This means
we can always tell if an object is constant.

This updates #20.