Parse method-calls properly..
This is a milestone, where I'll stop for the night.  We can now
parse "foo.bar()" into an identifier, followed by a second,
deliminated by a period.

To continue to allow old-style standard-library functions to work
we treat "directory.glob", etc, as exceptions.

However there is a break in compatiblity.  To parse things with
dots we now don't allow:

.3
.4

etc.

Instead floats must start with zero, for example "0.3", "0.4".