Updated handling of constant-values.
Rather than having a "const" flag associated with each object
we instead mark a variable as read-only in the environment.
This means then when constant-values are copied to new variables
their const-flag does not travel with them.

This still handles our scoping correctly, so that this works
as expected:

        const a = "Steve";

        function foo() {
                let a = "Meow";
                puts( "A is ", a, "\n" );
        };

        foo();

This closes #38.