Updated parser to fix some edge-cases.
Previously the parser would get into an infinite loop when
given an incomplete statement, such as:

    if ( true ) {
      let x = 1;

The parser would step forwards one character looking for the
mandatory closing `}` which didn't exist.

This pull-request adds sanity-checks to the `if`-handler, the
`let`, `const`, and `return` handlers.

I've added EOF/ILLEGAL token-prefix functions too, to get back
into a good state if we're broken.