Allow creating a temporary scope.
This is a bit horrid, but we allow creating a temporary scope
where ALL variables go to the parent EXCEPT some named values.

This is used for our `foreach` loop handler, to ensure that the
values we set for the duration of the block only last for the small
amount of time they should - and that any other changes in the block
work as users would expect.